crystalbet/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
pydantic==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.1.1
pymongo==4.4.1
bcrypt==4.0.1
python-dotenv==1.0.0
